Se modifica tipo de atributo en peso

+++ src/main/java/com/api/disney/entidades/Personaje.java
@@ -22,7 +22,7 @@ public class Personaje {
     private String imagen;
     private String nombre;
     private Integer edad;
-    private Integer peso;
+    private Double peso;
     private String historia;
     private Boolean alta;
